requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.6.0
